Padroniza AuditLog, limpa views monolíticas e setup/views
