+++ backend/apps/core/models/system_config.py
@@ -66,14 +66,14 @@ def audit_system_config_update(sender, instance, created, **kwargs):
     if old_valor == instance.valor:
         return
 
-    from .audit_log import AuditLog
+    from apps.core.services.audit import log_audit
 
-    AuditLog.objects.create(
+    log_audit(
         user=instance.atualizado_por,
-        acao="system_config_update",
+        acao="system_config_updated",
         modulo="core",
         entidade="SystemConfig",
-        entidade_id=str(instance.pk),
+        entidade_id=instance.pk,
         valores_anteriores={"valor": old_valor},
         valores_novos={"valor": instance.valor},
     )
