+++ backend/apps/core/views/__init__.py
@@ -1,0 +1,44 @@
+from .audit import AuditLogListView
+from .auth import (
+    LoginView,
+    LogoutView,
+    RefreshView,
+    logout_all,
+    me,
+    password_reset_confirm,
+    password_reset_request,
+)
+from .notifications import (
+    NotificationListView,
+    NotificationMarkReadView,
+    mark_all_read,
+    unread_count,
+)
+from .organizations import OrganizationViewSet
+from .system_config import SystemConfigDetailView, SystemConfigListView
+from .territorial import MunicipalityViewSet, RoleViewSet, StateViewSet, TerritoryViewSet
+from .users import UserViewSet
+
+
+__all__ = [
+    "AuditLogListView",
+    "LoginView",
+    "LogoutView",
+    "RefreshView",
+    "logout_all",
+    "me",
+    "password_reset_confirm",
+    "password_reset_request",
+    "NotificationListView",
+    "NotificationMarkReadView",
+    "mark_all_read",
+    "unread_count",
+    "OrganizationViewSet",
+    "SystemConfigDetailView",
+    "SystemConfigListView",
+    "MunicipalityViewSet",
+    "RoleViewSet",
+    "StateViewSet",
+    "TerritoryViewSet",
+    "UserViewSet",
+]
