+++ backend/setup/urls.py
@@ -17,28 +17,24 @@ Including another URLconf
 
 from django.contrib import admin
 from django.urls import include, path
-from rest_framework_simplejwt import views as jwt_views
-from rest_framework_simplejwt.views import TokenBlacklistView
 
-from .views import LoginView, RefreshView, LogoutView, logout_all, me, password_reset_request, password_reset_confirm
+from apps.core.views import (
+    LoginView,
+    RefreshView,
+    LogoutView,
+    logout_all,
+    me,
+    password_reset_request,
+    password_reset_confirm,
+)
 
 urlpatterns = [
     path("admin/", admin.site.urls),
-    path(
-        "api/v1/auth/login/",
-        LoginView.as_view(),
-        name="token_obtain_pair",
-    ),
-    path(
-        "api/v1/auth/token/refresh/",
-        RefreshView.as_view(),
-        name="token_refresh",
-    ),
+    path("api/v1/auth/login/", LoginView.as_view(), name="token_obtain_pair"),
+    path("api/v1/auth/token/refresh/", RefreshView.as_view(), name="token_refresh"),
     path("api/v1/auth/me/", me, name="me"),
     path("api/v1/", include("apps.core.urls")),
-    path("api/v1/auth/logout/", 
-         LogoutView.as_view(), 
-         name="token_blacklist"),
+    path("api/v1/auth/logout/", LogoutView.as_view(), name="token_blacklist"),
     path("api/v1/auth/logout-all/", logout_all, name="logout_all"),
     path("api/v1/auth/password-reset/request/", password_reset_request, name="password_reset_request"),
     path("api/v1/auth/password-reset/confirm/", password_reset_confirm, name="password_reset_confirm"),
